pyproject.toml
[project]
name = "b5"
version = "1.4.5"
description = "b5 - simple and sane task runner"
authors = [{ name = "David Danier", email = "danier@team23.de" }]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: BSD License",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Systems Administration",
]
requires-python = ">=3.10"
dependencies = [
    "Jinja2>=3.1,<4.0",
    "MarkupSafe>=3.0,<4.0",
    "PyYAML>=6.0,<7.0",
    "termcolor>=2.5,<3.0",
    "packaging>=24.0",
]

[project.urls]
Homepage = "https://github.com/team23/b5"
Repository = "https://github.com/team23/b5"
Documentation = "https://github.com/team23/b5/tree/master/docs"

[dependency-groups]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-cov>=6.0,<7.0",
    "ruff>=0.9.10",
]

[project.scripts]
b5 = "b5.main:main"
b5-init = "b5.init:main"
b5-execute = "b5.execute:main"

[tool.ruff]
line-length = 115
target-version = "py310"
output-format = "grouped"

[tool.ruff.lint]
select = ["F","E","W","C","I","N","UP","ANN","S","B","A","COM","C4","T20","PT","ARG","TD","RUF"]
ignore = ["A001","A002","A003","ANN401","C901","N8","B008","F405","F821","PT009"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"conftest.py" = ["S101","ANN","F401"]
"test_*.py" = ["S101","ANN","F401"]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "b5"
module-root = ""
